FIX: 캐릭터 추가
- 용병 42명, 몬스터 고블린 15마리

+++ Data/WaveData.cs
@@ -9,7 +9,7 @@ public class WaveData
     public Define.RaceType  race;               // 종족
     public int              hp;                 // 체력
     public int              defence;            // 방어력
-    public int              moveSpeed;          // 이동 속도
+    public float            moveSpeed;          // 이동 속도
     public int              gold;               // 골드
     public int              waveGold;           // 웨이브 성공 골드
     public int              maxEnemyCount;      // 적 생성 최대 횟수
